Holy Changes
Added Firebase dependency for ID Token verification
Added Authentication interceptor to verify ID Token and used RequestScopedData to pass the user UID along with the request.
Added FirebaseConfig.  This is where we initialize the firebase 'app' to allow us to do that ID Token verification.
Updated SimpleCORSFilter to allow for new header [Firebase-Auth].  You have to explicitly allow that since we aren't using * (wildcard)
New file in resources, serviceAccountkey.json.  That is for firebase as well.  I don't think we will want to keep that in the project as this is an open source project.  So we will need that to be moved but for now....whatevs.  We will need to regenerate the key file when we actually get this thing up and running.