Create Challenge updates
now saving the firebaseUuid passed in.  Updated DB table challenges column created_by and updated_by to be strings.  We are going to use the firebaseUuid as our primary indentifier.
Also removed the challenges/answer column.  Answers will now be held in the answers table.  This seemed the best approach for multiple choice questions.