Datasource updates
DatabaseConfig.java - Took out the now unnecessary NamedParameterJdbcTemplate and renamed the connection pool/datasource to match the project.

DailySheetServiceImpl.java - Switched to use @Inject.  It is still easily mocked out for testing purposes and I think it looks cleaner than the constructor injection.  I could be convinced of the constructor injection if Erik feels strongly about it.

application.properties- Referencing this stackoverflow.  The application startup time was nearly 23 seconds, used to be about 3 seconds.  It looks like this property is on by default but doesn't add anything for the Postgres environment.  If you look at the logs when the app is starting it shows an error and the LOB usage is turned off.